+++ src/routes/index.tsx
@@ -1,10 +1,9 @@
 import React from 'react';
-import { renderRoutes } from 'react-router-config';
-import { BrowserRouter } from 'react-router-dom';
-import staticRoutes from './staticRoutes';
+import { BrowserRouter, Route } from 'react-router-dom';
+// import staticRoutes from './staticRoutes';
 
 const AppRoutes: React.FC = () => {
-  return <BrowserRouter>{renderRoutes(staticRoutes)}</BrowserRouter>;
+  return <BrowserRouter>{}</BrowserRouter>;
 };
 
 export default AppRoutes;