+++ docs/CHANGELOG.md
@@ -36,4 +36,8 @@
 4. 路由
 * [react-router-dom]react的路由
 * [react-router-config]用于静态路由配置
-* [react-router] react-router-config的依耐
+* [react-router] react-router-config的依耐 -v6版本已经弃用
+
+
+二、 安装组件库
+* [antd组件库](https://ant.design/docs/react/introduce-cn)